+++ src/env.py
@@ -15,6 +15,8 @@ class RequiredEnv:
         self.ca_password = os.environ.get("CA_PASSWORD")
         self.grpc_port = os.environ.get("GRPC_PORT")
         self.connection_count = os.environ.get("CONNECTION_COUNT")
+        self.rabbitmq_url = os.environ.get("RABBITMQ_URL")
+        self.rabbitmq_exchange = os.environ.get("RABBITMQ_EXCHANGE")
 
         if self.deployment is None:
             raise Exception("Missing environment DEPLOYMENT")
@@ -28,3 +30,7 @@ class RequiredEnv:
             raise Exception("Missing environment GRPC_PORT")
         if self.connection_count is None:
             raise Exception("Missing environment CONNECTION_COUNT")
+        if self.rabbitmq_url is None:
+            raise Exception("Missing environment RABBITMQ_URL")
+        if self.rabbitmq_exchange is None:
+            raise Exception("Missing environment RABBITMQ_EXCHANGE")